djangoproject/requirements.txt
asgiref
coverage
Django
django-leaflet
djangorestframework
iniconfig
packaging
pluggy
psycopg2-binary
pytest
pytest-cov
pytest-django
pytz
sqlparse
typing_extensions
tzdata
django-import-export